Move search form elements and handlers to SearchForm
